Rewrite Api.js to use async/await Move Api.js to /util
